Move crm schema code to its own service
